More work on ItemPropertiesWidget.
